Label /sys/devices/system/cpu/cpufreq/ondemand and its files.
This directory and its files are created in sysfs when the
scaling_governor is set to "ondemand" by the init.<board>.rc file.
As this occurs after the initial restorecon_recursive("/sys") by init
and it does not appear to trigger any uevent notification and thus
will not be labeled by ueventd fixup_sys_perms(), we need to explicitly
invoke restorecon_recursive on it from the init.<board>.rc file after
setting the scaling_governor.

Change-Id: Iaf6e668f9ec499b92b4fcfe8b494bb79e25be3e5
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>