Move XTRA_SERVER entries to overlays
* XTRA_SERVERs are important, right? Like to get the almanac data
necessary for aGPS. Without the XTRA download, the chip needs to
collect sufficient navigation messages from the birds to compute
where they are in order to make sense of the ranging signals.
* Well, it seems that O doesn't like reading these entries from the
gps.conf file.

When in gps.conf:
GpsXtraDownloader: No XTRA servers were specified in the GPS
configuration

When in overlay:
<that noise doesn't exist>

* This seems to return GPS fix performance to what we had in N.

Change-Id: I82166036c6e4e4cc63aa87a0cb9e8ab5f11eccf5
Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Signed-off-by: Daniel Jarai <jaraidaniel@gmail.com>
Signed-off-by: Svyatoslav Ryhel <clamor95@gmail.com>
Signed-off-by: Sudokamikaze <sudokamikaze@protonmail.com>